Go back down to 0.2.x, but keeping our fixes
The 0.2.0.0 release on Hackage is flawed since its test modules aren't
included in the tarball.

We fixed this in 0.3.0.0, but that also changed the dependencies, which makes it
unbuildable in certain situations (e.g. with old nixpkgs sets).

This commit undoes those dependency changes, but keeps the properly-included
modules; hence it should be a working drop-in replacement for 0.2.0.0.

Note that there's still the issue of depending on lazysmallcheck2012, which
isn't on Hackage. I really like the lazysmallcheck2012 library, so it's a shame
to avoid using it; I'd rather see it uploaded to Hackage, but not sure if I can
take responsibility for that?

In any case, that would require changing the test suite, and I like the fact
that the code for panhandle has been working without change for years (all
recent commits just fiddle with the cabal and nix metadata).

If I ever have to actually change the code, I might consider dropping
lazysmallcheck2012. I might also consider bumping the pandoc JSON format!