Use IOHK's haskell.nix infrastructure
It's really simple to invoke, and seems to use Cabal to solve
dependencies, using pinned versions of Hackage. This means we don't have to
specify versions ourselves, and we get a bunch of nice features like useful
nix-shell integration, easy overrides, etc.

It also picks up external repos (like lazysmallcheck2012) from the cabal.project
file, which means we don't need to repeat ourselves specifying it for Nix and
Cabal separately.

Since it's so seamless, and doesn't need regenerating when our code changes,
I've made it a default.nix.